remove loadingCache and use card assetmanager exclusively
